User-Friendly updates to Sequence module (#3135)
