+++ src/Codeception/Module/Sequence.php
@@ -21,10 +21,9 @@ use Codeception\TestInterface;
  *
  * ``` php
  * <?php
- * 'post'.sq(1); // post_521fbc63021eb
- * 'post'.sq(2); // post_521fbc6302266
- * 'post'.sq(1); // post_521fbc63021eb
- * ?>
+ * sq('post1'); // post1_521fbc63021eb
+ * sq('post2'); // post2_521fbc6302266
+ * sq('post1'); // post1_521fbc63021eb
  * ```
  *
  * Example:
@@ -33,11 +32,10 @@ use Codeception\TestInterface;
  * <?php
  * $I->wantTo('create article');
  * $I->click('New Article');
- * $I->fillField('Title', 'Article'.sq('name'));
+ * $I->fillField('Title', sq('Article'));
  * $I->fillField('Body', 'Demo article with Lorem Ipsum');
  * $I->click('save');
- * $I->see('Article'.sq('name') ,'#articles')
- * ?>
+ * $I->see(sq('Article') ,'#articles')
  * ```
  *
  * Populating Database:
@@ -46,7 +44,7 @@ use Codeception\TestInterface;
  * <?php
  *
  * for ($i = 0; $i<10; $i++) {
- *      $I->haveInDatabase('users', array('login' => 'user'.sq($i), 'email' => 'user'.sq($i).'@email.com');
+ *      $I->haveInDatabase('users', array('login' => sq("user$i"), 'email' => sq("user$i").'@email.com');
  * }
  * ?>
  * ```
@@ -59,26 +57,58 @@ use Codeception\TestInterface;
  * {
  *     public function createUser(AcceptanceTester $I)
  *     {
- *         $I->createUser('email' . sqs('user') . '@mailserver.com', sqs('login'), sqs('pwd'));
+ *         $I->createUser(sqs('user') . '@mailserver.com', sqs('login'), sqs('pwd'));
  *     }
  *
  *     public function checkEmail(AcceptanceTester $I)
  *     {
- *         $I->seeInEmailTo('email' . sqs('user') . '@mailserver.com', sqs('login'));
+ *         $I->seeInEmailTo(sqs('user') . '@mailserver.com', sqs('login'));
  *     }
  *
  *     public function removeUser(AcceptanceTester $I)
  *     {
- *         $I->removeUser('email' . sqs('user') . '@mailserver.com');
+ *         $I->removeUser(sqs('user') . '@mailserver.com');
  *     }
  * }
  * ?>
+ * ```
+ *
+ * ### Config
+ *
+ * By default produces unique string with param as a prefix:
+ *
+ * ```
+ * sq('user') => 'user_876asd8as87a'
+ * ```
+ *
+ * This behavior can be configured using `prefix` config param.
+ *
+ * Old style sequences:
+ *
+ * ```yaml
+ * Sequence:
+ *     prefix: '_'
+ * ```
+ *
+ * Using id param inside prefix:
+ *
+ * ```yaml
+ * Sequence:
+ *     prefix: '{id}.'
  * ```
  */
 class Sequence extends CodeceptionModule
 {
     public static $hash = [];
     public static $suiteHash = [];
+    public static $prefix = '';
+
+    protected $config = ['prefix' => '{id}_'];
+
+    public function _initialize()
+    {
+        static::$prefix = $this->config['prefix'];
+    }
 
     public function _after(TestInterface $t)
     {
@@ -93,7 +123,6 @@ class Sequence extends CodeceptionModule
 
 if (!function_exists('sq') && !function_exists('sqs')) {
     require_once __DIR__ . '/../Util/sq.php';
-    require_once __DIR__ . '/../Util/sqs.php';
 } else {
     throw new ModuleException('Codeception\Module\Sequence', "function 'sq' and 'sqs' already defined");
 }
