+++ src/Codeception/Util/sq.php
@@ -7,11 +7,26 @@ if (!function_exists('sq')) {
         if ($id and isset(Sequence::$hash[$id])) {
             return Sequence::$hash[$id];
         }
-        $sequence = '_' . uniqid();
+        $prefix = str_replace('{id}', $id, Sequence::$prefix);
+        $sequence = $prefix . uniqid($id);
         if ($id) {
             Sequence::$hash[$id] = $sequence;
         }
         return $sequence;
     }
+}
 
+if (!function_exists('sqs')) {
+    function sqs($id = null)
+    {
+        if ($id and isset(Sequence::$suiteHash[$id])) {
+            return Sequence::$suiteHash[$id];
+        }
+        $prefix = str_replace('{id}', $id, Sequence::$prefix);
+        $sequence = $prefix . uniqid($id);
+        if ($id) {
+            Sequence::$suiteHash[$id] = $sequence;
+        }
+        return $sequence;
+    }
 }
